+++ packages/liveblocks-react/src/umbrella-store.ts
@@ -14,6 +14,7 @@ import type {
   Observable,
   OpaqueClient,
   Patchable,
+  Permission,
   Resolve,
   RoomNotificationSettings,
   Store,
@@ -533,6 +534,7 @@ export class SinglePageResource {
 
 type InternalState<M extends BaseMetadata> = Readonly<{
   optimisticUpdates: readonly OptimisticUpdate<M>[];
+  permissionsByRoom: Record<string, Set<Permission>>;
 
   // TODO: Ideally we would have a similar NotificationsDB, like we have ThreadDB
   notificationsById: Record<string, InboxNotificationData>;
@@ -646,6 +648,7 @@ export class UmbrellaStore<M extends BaseMetadata> {
     this._rawThreadsDB = new ThreadDB();
     this._store = createStore<InternalState<M>>({
       optimisticUpdates: [],
+      permissionsByRoom: {},
       notificationsById: {},
       settingsByRoomId: {},
       versionsByRoomId: {},
@@ -819,6 +822,10 @@ export class UmbrellaStore<M extends BaseMetadata> {
 
   public subscribe(callback: () => void): () => void {
     return this._store.subscribe(callback);
+  }
+
+  public _getPermissions(roomId: string): Set<Permission> | undefined {
+    return this._store.get().permissionsByRoom[roomId];
   }
 
   // Direct low-level cache mutations ------------------------------------------------- {{{
@@ -1293,21 +1300,41 @@ export class UmbrellaStore<M extends BaseMetadata> {
     return this._notifications.waitUntilLoaded();
   }
 
+  private updateRoomPermissions(permissions: Record<string, Permission[]>) {
+    const permissionsByRoom = { ...this._store.get().permissionsByRoom };
+
+    Object.entries(permissions).forEach(([roomId, newPermissions]) => {
+      // Get the existing set of permissions for the room and only ever add permission to this set
+      const existingPermissions = permissionsByRoom[roomId] ?? new Set();
+      // Add the new permissions to the set of existing permissions
+      newPermissions.forEach((permission) =>
+        existingPermissions.add(permission)
+      );
+      permissionsByRoom[roomId] = existingPermissions;
+    });
+
+    this._store.set((state) => ({
+      ...state,
+      permissionsByRoom,
+    }));
+  }
+
   public waitUntilRoomThreadsLoaded(
     roomId: string,
     query: ThreadsQuery<M> | undefined
   ) {
     const threadsFetcher = async (cursor?: string) => {
-      const room = this._client.getRoom(roomId);
-      if (room === null) {
-        throw new HttpError(`Room '${roomId}' is not available on client`, 479);
-      }
-
-      const result = await room.getThreads({ cursor, query });
+      const result = await this._client[kInternal].httpClient.getThreads({
+        roomId,
+        cursor,
+        query,
+      });
       this.updateThreadsAndNotifications(
         result.threads,
         result.inboxNotifications
       );
+
+      this.updateRoomPermissions(result.permissionHints);
 
       const lastRequestedAt =
         this._roomThreadsLastRequestedAtByRoom.get(roomId);
@@ -1354,12 +1381,8 @@ export class UmbrellaStore<M extends BaseMetadata> {
       return;
     }
 
-    const room = nn(
-      this._client.getRoom(roomId),
-      `Room with id ${roomId} is not available on client`
-    );
-
-    const updates = await room.getThreadsSince({
+    const updates = await this._client[kInternal].httpClient.getThreadsSince({
+      roomId,
       since: lastRequestedAt,
       signal,
     });
@@ -1371,6 +1394,8 @@ export class UmbrellaStore<M extends BaseMetadata> {
       updates.inboxNotifications.deleted
     );
 
+    this.updateRoomPermissions(updates.permissionHints);
+
     if (lastRequestedAt < updates.requestedAt) {
       // Update the `lastRequestedAt` value for the room to the timestamp returned by the current request
       this._roomThreadsLastRequestedAtByRoom.set(roomId, updates.requestedAt);
@@ -1381,7 +1406,9 @@ export class UmbrellaStore<M extends BaseMetadata> {
     const queryKey = makeUserThreadsQueryKey(query);
 
     const threadsFetcher = async (cursor?: string) => {
-      const result = await this._client[kInternal].getUserThreads_experimental({
+      const result = await this._client[
+        kInternal
+      ].httpClient.getUserThreads_experimental({
         cursor,
         query,
       });
@@ -1389,6 +1416,8 @@ export class UmbrellaStore<M extends BaseMetadata> {
         result.threads,
         result.inboxNotifications
       );
+
+      this.updateRoomPermissions(result.permissionHints);
 
       // We initialize the `_userThreadsLastRequestedAt` date using the server timestamp after we've loaded the first page of inbox notifications.
       if (this._userThreadsLastRequestedAt === null) {
@@ -1422,7 +1451,7 @@ export class UmbrellaStore<M extends BaseMetadata> {
 
     const result = await this._client[
       kInternal
-    ].getUserThreadsSince_experimental({
+    ].httpClient.getUserThreadsSince_experimental({
       since: lastRequestedAt,
       signal,
     });
@@ -1437,6 +1466,8 @@ export class UmbrellaStore<M extends BaseMetadata> {
       result.threads.deleted,
       result.inboxNotifications.deleted
     );
+
+    this.updateRoomPermissions(result.permissionHints);
   }
 
   public waitUntilRoomVersionsLoaded(roomId: string) {