+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.12.1
+
+### `@liveblocks/react-ui`
+
+- Prevent unsupported attachment previews from loading infinitely.
+- Refactored `Thread` and `Comment` component to be used outside of the
+  `RoomProvider` component.
+
 ## 2.12.0
 
 This release adds support for tracking synchronization status of pending local